# Remove gpu_kernel_with_index