--- a/aten/src/ATen/native/cuda/RangeFactories.cu
+++ b/aten/src/ATen/native/cuda/RangeFactories.cu
@@ -26,7 +26,7 @@ __global__ void elementwise_kernel_with_index(index_t N, func_t f, typename func
 }
 
 template<typename index_t, typename func_t>
-void gpu_kernel_with_index(Tensor &output, f) {
+void gpu_kernel_with_index(Tensor &output, func_t f) {
   int64_t N = output.numel();
   int64_t grid = (N + block_work_size - 1) / block_work_size;
   auto stream = at::cuda::getCurrentCUDAStream();