@ngimel 2020-02-15
Test failure in test_print and test_pickle might be related?