@ngimel 2020-02-15
This generally looks good, but if you want to be awesome, can you implement @largeTensorTest wrapper in common_device_type.py? It seems to be usefule, TEST_LARGE_TENSOR is used in a lot of places.