@zasdfgbnm 2020-02-15
@ngimel The failures are related. They are about n dimensional zero sized tensors.