@ngimel 2020-02-15
test_nn now segfaults.