@zasdfgbnm 2020-02-15
@ngimel 
```
CUDA out of memory. Tried to allocate 6.00 GiB (GPU 0; 14.85 GiB total capacity; 8.00 GiB already allocated; 1.96 GiB free; 12.13 GiB reserved in total by PyTorch)
```
It is strange that this test is not skipped on CI. It was correctly skipped locally on my machine. Could it be possible that the total RAM of GPU is 32GB, but only 16GB is allowed by PyTorch to use?