@ngimel 2020-02-16
Which job are you looking at? Generally in `Set up CI environment` section you can see output of `nvidia-smi`. E.g. here https://app.circleci.com/jobs/github/pytorch/pytorch/4510128 nvidia-smi reports M60 with 8 GB memory, and the test fails with the correct report (tried to allocate 4 GB, 7.44 or something available), so it's not being skipped correctly?
non-gpu tests fail with segfault, not with oom.