@zasdfgbnm 2020-02-16
@ngimel Yes, you are right, it is not skipped correctly. I was looking at a build on a wrong directory, in which it was successfully skipped, that's why I was saying
```
It is strange that this test is not skipped on CI. It was correctly skipped locally on my machine.
```
The SIGSEGV fault could be because some test was skipped both on CPU and CUDA (due to the `skipIf`), but now it is only skipped on CUDA. But the machine on the CI does not have that much of host memory to run it on CPU. Let me check on a 256GB RAM computer.